backend/requirements.txt
Flask
pymongo
flask-cors
python-dotenv
bcrypt
PyJWT
